Move enums because typecheck complains
